Move constraints to separate module
